Rename `workouts-table-dialog-bar-types.ts` file to `types.ts` and update imports
